cleared up handlers/handler naming removed unused server configs removed check_suite code unrelated to this PR
